requirements.txt
2selenium==4.9.0
python-dotenv==1.0.0